--- app/src/main/java/com/example/chapmac/rakkan/mqtt_app_test/Adapter.java
@@ -1,95 +1,0 @@
-package com.example.chapmac.rakkan.mqtt_app_test;
-
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import com.example.chapmac.rakkan.mqtt_app_test.Subscribe.SubscribeItem;
-
-import java.util.ArrayList;
-
-public class Adapter extends RecyclerView.Adapter<Adapter.ViewHolder> {
-
-    private ArrayList<SubscribeItem> subscribeList;
-    private OnItemClickListener mListener;
-
-    public interface OnItemClickListener{
-        void onItemClick(int position);
-        void onDeleteClick(int position);
-    }
-
-    public void setOnCilckItemListener(OnItemClickListener listener){
-        mListener = listener;
-    }
-
-    public static class ViewHolder  extends RecyclerView.ViewHolder{
-
-        public ImageView mImageView;
-        public TextView mTextView1;
-        public TextView mTextView2;
-        public ImageView mDeleteImage;
-
-        public ViewHolder(View itemView, final OnItemClickListener listener){
-            super(itemView);
-            mImageView = itemView.findViewById(R.id.imageView);
-            mTextView1 = itemView.findViewById(R.id.textView);
-            mTextView2 = itemView.findViewById(R.id.textView1);
-            mDeleteImage = itemView.findViewById(R.id.image_delete);
-
-            itemView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    if(listener != null){
-                        int position = getAdapterPosition();
-                        if(position != RecyclerView.NO_POSITION){
-                            listener.onItemClick(position);
-                        }
-                    }
-                }
-            });
-
-            mDeleteImage.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    if(listener != null){
-                        int position = getAdapterPosition();
-                        if(position != RecyclerView.NO_POSITION){
-                            listener.onDeleteClick(position);
-                        }
-                    }
-                }
-            });
-        }
-    }
-
-    public Adapter(ArrayList<SubscribeItem> subscribeItems){
-        subscribeList = subscribeItems;
-    }
-
-    @NonNull
-    @Override
-    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.subscribe_item,parent,false);
-        ViewHolder viewHolder = new ViewHolder(view,mListener);
-        return viewHolder;
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        SubscribeItem currentItem = subscribeList.get(position);
-
-        holder.mImageView.setImageResource(currentItem.getmImageResource());
-        holder.mTextView1.setText(currentItem.getMtext1());
-        holder.mTextView2.setText(currentItem.getMtext2());
-    }
-
-    @Override
-    public int getItemCount() {
-        return subscribeList.size();
-    }
-
-}
