+++ app/src/main/java/com/example/chapmac/rakkan/mqtt_app_test/Subscribe/SubscribeAdapter.java
@@ -1,0 +1,53 @@
+package com.example.chapmac.rakkan.mqtt_app_test.Subscribe;
+
+import android.support.annotation.NonNull;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.chapmac.rakkan.mqtt_app_test.R;
+
+import java.util.ArrayList;
+
+public class SubscribeAdapter extends RecyclerView.Adapter<SubscribeViewHolder> {
+
+    private ArrayList<SubscribeItem> subscribeList;
+    private OnItemClickListener mListener;
+
+    public interface OnItemClickListener{
+        void onItemClick(int position);
+        void onDeleteClick(int position);
+    }
+
+    public void setOnCilckItemListener(OnItemClickListener listener){
+        mListener = listener;
+    }
+
+    public SubscribeAdapter(ArrayList<SubscribeItem> subscribeItems){
+        subscribeList = subscribeItems;
+    }
+
+    @NonNull
+    @Override
+    public SubscribeViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.subscribe_item,parent,false);
+        SubscribeViewHolder subscribeViewHolder = new SubscribeViewHolder(view,mListener);
+        return subscribeViewHolder;
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull SubscribeViewHolder holder, int position) {
+        SubscribeItem currentItem = subscribeList.get(position);
+
+        holder.mImageView.setImageResource(currentItem.getmImageResource());
+        holder.mTextView1.setText(currentItem.getMtext1());
+        holder.mTextView2.setText(currentItem.getMtext2());
+    }
+
+    @Override
+    public int getItemCount() {
+        return subscribeList.size();
+    }
+
+}
